NFASimulator.java
Reads the definition of the machine from a file (the first command line argument), with
the second command line argument being the string to simulate the machine running on (the input to the
automata). The output  is written to standard output. The output consists of
either: 1) the word accept followed by a blank space followed by the list of accept states (blank space
delimited) that the automata can end up in after reading in the input string (if there is a way for the
automata to end in an accept state after reading the input); or 2) the word reject followed by a blank
space followed by the list of states (blank space delimited) that automata can end up in after reading the
string (if there is no way for the automata to end in an accept state after reading the input). 

There are two types of input lines: state lines and transition lines.
The state lines are of the form:
state x <y> <z>
state x <z> <y>
where x is a number in [0, 1000], y is the text start, and z is the text accept. Both <y> and <z> are
optional.

The format of the transitions in the file is:
transition    p   x   q
where p and q are states in [0, 1000], and x is the symbol read.